add Validate and Setup tasks
These tasks are used to setup a deploy.  They validate the
params and create the deploy dirs and clone the A/B repos.

I chose to do the validation logic as a callback task as I need
to also use it as a callback for the coming Update and Link tasks.