add Update and Link tasks
These tasks update the deploy release repo's source and
links the deploy release dir and the current dir respectively.
These are more steps needed to rollout source in the A/B deploy
scheme.

The Update task first reads what the current link point to and
then sets the current release dir param to that value.  It then
sets the other release dir as the deploy release dir param. Once
the deploy release dir has been determined, the task then runs a
cmd to hard reset the git repo to the given ref param value.

The Link task is designed to be run at some point after the Update
task has run.  It takes deploy release dir param set by the Update
task and links it as the current dir.  This allows everything that
relies on the source to reference via this current link, which is
updated on deploy.

I also made of few changes to the setup and validate tasks:

* setup: switched to `ssh!` for making the ssh cmds so they'd
  raise an exception if they fail.
* setup: renamed the private cmd string method to be more
  specific as to what it returns.
* setup: removed the pathname require as it isn't needed.
* validate: added setting the current dir param (which is needed
  by the link task).