add BundlerCmdStr util module
This is a helper for building bundler cmd strings.  This builds a
cmd string that cd's to a root path (the current release dir by
default) and bundle exec's a given cmd string.  This also includes
an Assert-style test helper for testing BundlerCmdStrs.

Note: this is a copy of the cap-style version from the CapUtil
gem.