add CurrentGitBranch util module
This is a helper for looking up the current git branch in deploy
scripts.  If no block is given, this just returns a cmd str that
can be run that will return the raw info on stdout.  If a block
is given, it yields the cmd string and expects the block will
execute the cmd string and return the local cmd object.  It then
parses the stdout and returns the current branch name.

I chose to yield to a given block for running the cmd string.
This allows the user to control how the cmd is run (`cmd` vs
`cmd!`) and also allows the user to add any custom log message.
It also simplifies the API as I don't have to have some task
passed in and then instance eval on it.

Note: this is a rework of the cap-style version from the CapUtil
gem.