add a test helper for setting Validate task params/ivars
I discovered, when testing a lib that used dk-abdeploy, that I
needed the params to be in the state the validate task would leave
them in.  This is a public test helper that sets up the params and
also sets up associated ivars to use in tests.  The validate
support file here has been updated to use this.  This allows other
libs to get the same validate test support.