rework the cmd to readlink the current release as an ssh cmd
Ssh cmds now have stdout attrs.  This value will be the first non
empty output from the hosts the ssh cmd runs on.  In this case,
we are only running the cmd on 1 host so it is identical to
manually building the ssh cmd like we did before.

This also updates the logic to default to the 'B' release dir if
the readlink stdout is empty.  This is a valid case - it happens
if this is the initial deploy.  This means that initial deploys
will always deploy to the A release dir (as was the original
intent).

Finally, this makes the readlink cmd str a class method like the
git reset cmd str so it can be used for stubbing, etc.

I noticed all this b/c I was trying to use the new cleanup task
and it kept saying the current release dir was empty.  That was
b/c this updated task would never set a value on the initial
deploy.  This fixes that issue.